Add open library client
https://github.com/rubyforgood/literacy-network/issues/34

Open library provides us with important book info we need when adding a
book to the inventory system. We need to be able to find book info by
ISBN.

OpenLibrary asks that we not hit the API too much. We should look at
adding some kind of caching in the future. To make progress on lowering
traffic, we are using long lived HTTP connection. Because we are running
in a threaded environment, we need to wrap the connections in a thread
safe connection pool. More work will be needed to repair broken
connections, but his is a good start.